requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# STT
openai-whisper==20231117
openai==1.30.1

# 운율 분석
librosa==0.10.2
numpy==1.26.4
scipy==1.13.0
fastdtw==0.3.4

# AI 대화
google-generativeai==0.5.4

# 오디오 처리
soundfile==0.12.1
pydub==0.25.1
